# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stealth Console API
  version: 0.1.0
  description: Control-plane API for Stealth Console. Browser sessions are opaque HttpOnly cookies and are never returned in JSON.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/organizations/{organizationID}/traces:
    get:
      operationId: listOrganizationTraces
      description: List the bounded durable root-request trace index visible to organization members. Nested spans and full OpenTelemetry attributes remain in the private telemetry backend.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/OrganizationID' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' }]
      responses: { '200': { description: Durable organization request traces, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationTracesPage' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' }, '500': { $ref: '#/components/responses/InternalError' } }
...
  /v1/projects/{projectID}/usage:
    get:
      operationId: getProjectUsage
      description: Return a live aggregate snapshot of tenant-owned project resources. Network egress, compute time, and invoice totals are intentionally omitted until durable metering exists.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }]
      responses:
        '200': { description: Project usage snapshot, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectUsageResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /v1/projects/{projectID}/sites/{siteID}/domains:
    get:
      operationId: listSiteDomains
      description: List custom hostnames bound to a Site. API-key callers require sites.read.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/SiteID' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' }]
      responses: { '200': { description: Site domains, content: { application/json: { schema: { $ref: '#/components/schemas/SiteDomainsPage' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
    post:
      operationId: createSiteDomain
      description: Bind a DNS hostname to a Site and return the TXT challenge record. The hostname serves only after verification.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/SiteID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateSiteDomainRequest' } } } }
      responses: { '201': { description: Domain binding created, content: { application/json: { schema: { $ref: '#/components/schemas/SiteDomainResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /v1/projects/{projectID}/sites/{siteID}/domains/{domainID}:
    get:
      operationId: getSiteDomain
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/SiteID' }, { $ref: '#/components/parameters/DomainID' }]
      responses: { '200': { description: Site domain, content: { application/json: { schema: { $ref: '#/components/schemas/SiteDomainResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      operationId: deleteSiteDomain
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/SiteID' }, { $ref: '#/components/parameters/DomainID' }]
      responses: { '204': { description: Domain binding deleted }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /v1/projects/{projectID}/sites/{siteID}/domains/{domainID}/verify:
    post:
      operationId: verifySiteDomain
      description: Verify the required TXT record at _stealth-verification.hostname. Verified hosts may use the optional in-process ACME manager or an external TLS terminator.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/SiteID' }, { $ref: '#/components/parameters/DomainID' }]
      responses: { '200': { description: Verified domain, content: { application/json: { schema: { $ref: '#/components/schemas/SiteDomainResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: stealth_session }
    projectAppCookie: { type: apiKey, in: cookie, name: stealth_app_<project-id-without-hyphens>, description: Project-specific opaque HttpOnly cookie. The actual name is derived from the path projectID; it is never the Console stealth_session cookie. }
    projectApiKey: { type: apiKey, in: header, name: X-Stealth-Key, description: Project-bound server-to-server secret. Supported scopes are users.read/write, databases.read/write, storage.read/write, functions.read/write, sites.read/write, webhooks.read/write, and realtime.read; never send this header through the Console browser bridge. }
  parameters:
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Cursor: { name: cursor, in: query, schema: { type: string, format: uuid } }
    OrganizationID: { name: organizationID, in: path, required: true, schema: { type: string, format: uuid } }
...
    ProjectID: { name: projectID, in: path, required: true, schema: { type: string, format: uuid } }
...
    SiteID: { name: siteID, in: path, required: true, schema: { type: string, format: uuid } }
...
    DomainID: { name: domainID, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    BadRequest: { description: Invalid JSON or parameter, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    Forbidden: { description: Insufficient organization role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Conflict: { description: Resource conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    InternalError: { description: Internal server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
...
    Pagination: { type: object, required: [limit, next_cursor], properties: { limit: { type: integer }, next_cursor: { type: [string, 'null'], format: uuid } } }
    ErrorEnvelope: { type: object, additionalProperties: false, required: [error], properties: { error: { type: object, additionalProperties: false, required: [code, message], properties: { code: { type: string }, message: { type: string } } } } }
...
    ProjectUsage: { type: object, additionalProperties: false, required: [project_id, captured_at, application_users, database_count, database_table_count, database_row_count, storage_file_count, storage_bytes, storage_quota_bytes, function_count, function_artifact_bytes, function_quota_bytes, site_count, site_artifact_bytes, site_reserved_bytes, site_quota_bytes, realtime_event_count, webhook_delivery_count_7d], properties: { project_id: { type: string, format: uuid }, captured_at: { type: string, format: date-time }, application_users: { type: integer, format: int64, minimum: 0 }, database_count: { type: integer, format: int64, minimum: 0 }, database_table_count: { type: integer, format: int64, minimum: 0 }, database_row_count: { type: integer, format: int64, minimum: 0 }, storage_file_count: { type: integer, format: int64, minimum: 0 }, storage_bytes: { type: integer, format: int64, minimum: 0 }, storage_quota_bytes: { type: integer, format: int64, minimum: 0 }, function_count: { type: integer, format: int64, minimum: 0 }, function_artifact_bytes: { type: integer, format: int64, minimum: 0 }, function_quota_bytes: { type: integer, format: int64, minimum: 0 }, site_count: { type: integer, format: int64, minimum: 0 }, site_artifact_bytes: { type: integer, format: int64, minimum: 0 }, site_reserved_bytes: { type: integer, format: int64, minimum: 0 }, site_quota_bytes: { type: integer, format: int64, minimum: 0 }, realtime_event_count: { type: integer, format: int64, minimum: 0 }, webhook_delivery_count_7d: { type: integer, format: int64, minimum: 0 } } }
    ProjectUsageResponse: { type: object, additionalProperties: false, required: [usage], properties: { usage: { $ref: '#/components/schemas/ProjectUsage' } } }
...
    HTTPTrace: { type: object, additionalProperties: false, required: [id, trace_id, service, method, route, status, duration_ms, response_bytes, started_at, finished_at, created_at], properties: { id: { type: string, format: uuid }, trace_id: { type: string, minLength: 16, maxLength: 64, pattern: '^[0-9a-fA-F]+$' }, span_id: { type: [string, 'null'], minLength: 8, maxLength: 32, pattern: '^[0-9a-fA-F]+$' }, organization_id: { type: [string, 'null'], format: uuid }, project_id: { type: [string, 'null'], format: uuid }, organization_name: { type: string, maxLength: 255 }, project_name: { type: string, maxLength: 255 }, service: { type: string, minLength: 1, maxLength: 64 }, method: { type: string, minLength: 3, maxLength: 16 }, route: { type: string, minLength: 1, maxLength: 240 }, status: { type: integer, minimum: 100, maximum: 599 }, duration_ms: { type: integer, format: int64, minimum: 0 }, response_bytes: { type: integer, format: int64, minimum: 0 }, started_at: { type: string, format: date-time }, finished_at: { type: string, format: date-time }, created_at: { type: string, format: date-time } } }
    OrganizationTracesPage: { type: object, additionalProperties: false, required: [traces, pagination], properties: { traces: { type: array, items: { $ref: '#/components/schemas/HTTPTrace' } }, pagination: { $ref: '#/components/schemas/Pagination' } } }
...
    SiteDomain: { type: object, additionalProperties: false, required: [id, project_id, site_id, hostname, status, verification_token, verification_record_name, verification_record_type, verification_record_value, tls_status, created_at, updated_at], properties: { id: { type: string, format: uuid }, project_id: { type: string, format: uuid }, site_id: { type: string, format: uuid }, hostname: { type: string, pattern: '^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?(?:\.[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?)+$' }, status: { type: string, enum: [pending, verified, disabled] }, verification_token: { type: string, pattern: '^[0-9a-f]{64}$' }, verification_record_name: { type: string }, verification_record_type: { type: string, enum: [TXT] }, verification_record_value: { type: string }, verified_at: { type: [string, 'null'], format: date-time }, tls_status: { type: string, enum: [external, pending, active, failed] }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } } }
    CreateSiteDomainRequest: { type: object, additionalProperties: false, required: [hostname], properties: { hostname: { type: string, minLength: 4, maxLength: 253, description: Lowercase DNS hostname; IP addresses and wildcard names are not supported. } } }
...
    SiteDomainResponse: { type: object, additionalProperties: false, required: [domain], properties: { domain: { $ref: '#/components/schemas/SiteDomain' } } }
    SiteDomainsPage: { type: object, additionalProperties: false, required: [domains, pagination, can_manage], properties: { domains: { type: array, items: { $ref: '#/components/schemas/SiteDomain' } }, pagination: { $ref: '#/components/schemas/Pagination' }, can_manage: { type: boolean } } }
...